backend/pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = "AI-powered quantitative trading bot with live news detection and predictive algorithms"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "websockets>=14.0",
    # Async HTTP
    "httpx>=0.28.0",
    "aiohttp>=3.11.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.21.0",
    "alembic>=1.14.0",
    # Broker SDKs
    "alpaca-py>=0.35.0",
    # Data providers
    "finnhub-python>=2.4.20",
    # Data analysis
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    # Technical indicators
    "ta>=0.11.0",
    # ML
    "scikit-learn>=1.6.0",
    "xgboost>=2.1.0",
    "joblib>=1.4.0",
    # Sentiment (FinBERT)
    "transformers>=4.47.0",
    "torch>=2.5.0",
    # Scheduling
    "apscheduler>=3.10.0",
    # Utils
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
